requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Crawl4AI==0.5.0.post4
cryptography==44.0.2
cssselect==1.3.0
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.1.0
faust-cchardet==2.1.19
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
googlesearch-python==1.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.12.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.20
langchain-core==0.3.44
langchain-text-splitters==0.3.6
langsmith==0.3.13
litellm==1.63.11
llama-cloud==0.1.14
llama-cloud-services==0.6.5
llama-index==0.12.24
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.24.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-ollama==0.5.3
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
ollama==0.4.7
openai==1.66.3
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
playwright==1.50.0
propcache==0.3.0
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pypdf==5.3.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.39
striprtf==0.0.26
tenacity==9.0.0
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
faiss-cpu==1.7.4
